+++ enclone_print/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod define_mat;
 pub mod filter;
 pub mod finish_table;
+pub mod gene_scan;
 pub mod loupe;
 pub mod print_clonotypes;
 pub mod print_utils1;
